linting

+++ projects/hypermodel/data.py
@@ -7,9 +7,11 @@ from utils import *
 import sklearn.gaussian_process.kernels
 from sklearn.gaussian_process.kernels import RBF
 
+
 def polynomial(coeffs):
     def f(x):
         return jnp.polyval(jnp.array(coeffs), x)
+
     return f
 
 
@@ -20,23 +22,21 @@ def generate_points(poly_fun, rng_key, xlim=[-1, 1], num_points=100):
     return {'x': x_points, 'y': y_points}
 
 
-
 # Copied from Robert's Perceiver repo
 # Reference:
 # https://kazemnejad.com/blog/transformer_architecture_positional_encoding/
 
 def fourier_positional_encoding(x, max_freq, num_bands, base):
-    x = jnp.expand_dims(x,-1)
+    x = jnp.expand_dims(x, -1)
     dtype, orig_x = x.dtype, x
 
-    scales = jnp.logspace(0., jnp.log(max_freq / 2) / jnp.log(base), num_bands, base = base, dtype = dtype)
+    scales = jnp.logspace(0., jnp.log(max_freq / 2) / jnp.log(base), num_bands, base=base, dtype=dtype)
     scales = scales[(*((None,) * (len(x.shape) - 1)), Ellipsis)]
 
     x = x * scales * jnp.pi
-    x = jnp.concatenate([jnp.sin(x), jnp.cos(x)], axis =-1)
-    x = jnp.concatenate((x, orig_x), axis = -1)
+    x = jnp.concatenate([jnp.sin(x), jnp.cos(x)], axis=-1)
+    x = jnp.concatenate((x, orig_x), axis=-1)
     return x
-
 
 
 class PolyDataset(Dataset):
@@ -52,7 +52,7 @@ class PolyDataset(Dataset):
 
 
 def create_train_test_loaders(x, y, train_split=0.8, batch_size=64):
-    num_train = int(train_split*len(x))
+    num_train = int(train_split * len(x))
 
     train_x = x[:num_train]
     train_y = y[:num_train]
@@ -62,12 +62,13 @@ def create_train_test_loaders(x, y, train_split=0.8, batch_size=64):
     test_y = y[num_train:]
     test_ds = PolyDataset(test_x, test_y)
 
-    return DataLoader(train_ds, batch_size=batch_size, shuffle=True), DataLoader(test_ds, batch_size=batch_size, shuffle=False)
+    return DataLoader(train_ds, batch_size=batch_size, shuffle=True), DataLoader(test_ds, batch_size=batch_size,
+                                                                                 shuffle=False)
 
 
 def get_poly_data(config):
     coeff = omegaconf_list_to_array(config.data.polynomial.coeff)
-    coefficients = jnp.array(coeff)*config.data.polynomial.scale
+    coefficients = jnp.array(coeff) * config.data.polynomial.scale
     f = polynomial(coefficients)
     key = jax.random.PRNGKey(0)
     points = generate_points(f, key, xlim=config.data.xlim, num_points=config.data.num_points)
@@ -76,10 +77,8 @@ def get_poly_data(config):
     params = config.data.fourier
     encoded_points = encode(points['x'], params.max_freq, params.num_bands, params.base)
 
-    return create_train_test_loaders(encoded_points, points['y'], 
-                train_split=config.data.train_split, batch_size=config.train.batch_size)
-
-
+    return create_train_test_loaders(encoded_points, points['y'],
+                                     train_split=config.data.train_split, batch_size=config.train.batch_size)
 
 
 def gaussian_process(key, num_draw_points, num_train_points, xlims, l):
@@ -100,6 +99,4 @@ def gaussian_process(key, num_draw_points, num_train_points, xlims, l):
     y_train = y_vals[num_draw_points:]
     x_draw = all_points[:num_draw_points]
     y_draw = y_vals[:num_draw_points]
-    return x_train, y_train, x_draw, y_draw 
-
-
+    return x_train, y_train, x_draw, y_draw
